Hack/fix: ignore old LogEntries in Log
It is something that I think should be sanitized in the hub, not sure
why it's happening thought so further investigation is needed. I'm in
a process of removing the old sockets for transport::socket so would
prefer to finish that to (I don't believe this bug is related to
transport though).

For now simply ignoring the message seem to do the trick.